+++ src/main/java/io/dsub/discogs/api/artist/dto/ArtistDTO.java
@@ -3,4 +3,4 @@ package io.dsub.discogs.api.artist.dto;
 import lombok.Builder;
 
 @Builder
-public record ArtistDTO (Integer id, String name, String realName, String profile, String dataQuality) {}
+public record ArtistDTO (Long id, String name, String realName, String profile, String dataQuality) {}
