+++ src/main/java/io/dsub/discogs/api/label/service/LabelService.java
@@ -2,8 +2,8 @@ package io.dsub.discogs.api.label.service;
 
 import io.dsub.discogs.api.label.command.LabelCommand;
 import io.dsub.discogs.api.label.dto.LabelDTO;
-import io.dsub.discogs.api.service.PagingService;
-import io.dsub.discogs.api.service.ValidatingService;
+import io.dsub.discogs.api.core.service.PagingService;
+import io.dsub.discogs.api.core.service.ValidatingService;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import reactor.core.publisher.Flux;
@@ -12,9 +12,8 @@ import reactor.core.publisher.Mono;
 public interface LabelService extends ValidatingService, PagingService {
     Flux<LabelDTO> getLabels();
     Mono<Page<LabelDTO>> getLabels(Pageable pageable);
-    Mono<LabelDTO> updateLabel(Integer id, LabelCommand.Update command);
+    Mono<LabelDTO> updateLabel(long id, LabelCommand.Update command);
     Mono<LabelDTO> saveOrUpdate(LabelCommand.Create command);
-    Mono<Void> deleteLabel(Integer id);
-
-    Mono<LabelDTO> findById(Integer id);
+    Mono<Void> deleteLabel(long id);
+    Mono<LabelDTO> findById(long id);
 }
